+++ SharedContracts/SharedContracts/Events/Orders/OrderProcessedEvent.cs
@@ -19,13 +19,13 @@ public class OrderProcessedEvent
     public decimal TotalAmount { get; set; }
     public decimal DeliveryFee { get; set; }
 
-    public decimal DeliveryLatitude { get; set; }
-    public decimal DeliveryLongitude { get; set; }
+    public double DeliveryLatitude { get; set; }
+    public double DeliveryLongitude { get; set; }
     public string DeliveryAddress { get; set; } = string.Empty;
     public string DeliveryCity { get; set; } = string.Empty;
 
-    public decimal RestaurantLatitude { get; set; }
-    public decimal RestaurantLongitude { get; set; }
+    public double RestaurantLatitude { get; set; }
+    public double RestaurantLongitude { get; set; }
     public string RestaurantAddress { get; set; } = string.Empty;
     public string RestaurantCity { get; set; } = string.Empty;
